services/ingestion_service/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ingestion_service"
version = "0.1.0"
description = "Bibliotalk ingestion service"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "PyYAML>=6.0.1",
  "rich>=13.7.0",
  "tenacity>=8.5.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.8",
]
ingest = [
  "youtube-transcript-api>=0.6.2",
]

[tool.setuptools]
packages = [
  "ingestion_service",
  "ingestion_service.adapters",
  "ingestion_service.domain",
  "ingestion_service.pipeline",
  "ingestion_service.runtime",
]

[tool.setuptools.package-dir]
"ingestion_service" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
